+++ core-ng/src/main/java/core/framework/http/HTTPRequest.java
@@ -10,26 +10,6 @@ import java.util.Map;
  * @author neo
  */
 public final class HTTPRequest {
-    public static HTTPRequest get(String uri) {
-        return new HTTPRequest(HTTPMethod.GET, uri);
-    }
-
-    public static HTTPRequest post(String uri) {
-        return new HTTPRequest(HTTPMethod.POST, uri);
-    }
-
-    public static HTTPRequest put(String uri) {
-        return new HTTPRequest(HTTPMethod.PUT, uri);
-    }
-
-    public static HTTPRequest delete(String uri) {
-        return new HTTPRequest(HTTPMethod.DELETE, uri);
-    }
-
-    public static HTTPRequest patch(String uri) {
-        return new HTTPRequest(HTTPMethod.PATCH, uri);
-    }
-
     private final String uri;
     private final HTTPMethod method;
     private final Map<String, String> headers = Maps.newLinkedHashMap();