+++ core-ng/src/main/java/core/framework/module/RouteConfig.java
@@ -2,6 +2,8 @@ package core.framework.module;
 
 import core.framework.http.HTTPMethod;
 import core.framework.impl.module.ModuleContext;
+import core.framework.impl.web.HTTPServerHealthCheckHandler;
+import core.framework.util.Exceptions;
 import core.framework.web.Controller;
 
 /**
@@ -35,6 +37,7 @@ public final class RouteConfig {
     }
 
     public void add(HTTPMethod method, String path, Controller controller) {
+        if (HTTPServerHealthCheckHandler.PATH.equals(path)) throw Exceptions.error("health-check path is reserved by framework, path={}", path);
         context.route(method, path, controller, false);
     }
 }